Move Layer and OCRConfidenceLayer classes from layer.py to annotations.py
